Move news/Read endpoints to collections/
