+++ src/views/news/show.phtml
@@ -185,7 +185,7 @@
                 </summary>
 
                 <nav class="popup__container popup__container--bottom">
-                    <form method="post" action="<?= url('mark news as read') ?>">
+                    <form method="post" action="<?= url('mark collection as read', ['id' => $news->id]) ?>">
                         <input type="hidden" name="csrf" value="<?= $csrf_token ?>" />
 
                         <button class="popup__item popup__item--button">
@@ -194,7 +194,7 @@
                         </button>
                     </form>
 
-                    <form method="post" action="<?= url('read news later') ?>">
+                    <form method="post" action="<?= url('read collection later', ['id' => $news->id]) ?>">
                         <input type="hidden" name="csrf" value="<?= $csrf_token ?>" />
 
                         <button class="popup__item popup__item--button">
@@ -207,7 +207,7 @@
 
                     <form
                         method="post"
-                        action="<?= url('mark news to never read') ?>"
+                        action="<?= url('never read collection', ['id' => $news->id]) ?>"
                         data-controller="confirmation"
                         data-action="confirmation#confirm"
                         data-confirmation-message-value="<?= _('You’ll remove all the news links, this action cannot be canceled. Are you sure?') ?>"
