+++ tests/controllers/collections/ReadTest.php
@@ -6,11 +6,12 @@ use flusio\models;
 
 class ReadTest extends \PHPUnit\Framework\TestCase
 {
-    use \tests\LoginHelper;
     use \tests\FakerHelper;
-    use \Minz\Tests\FactoriesHelper;
+    use \tests\FlashAsserts;
     use \tests\InitializerHelper;
+    use \tests\LoginHelper;
     use \Minz\Tests\ApplicationHelper;
+    use \Minz\Tests\FactoriesHelper;
     use \Minz\Tests\ResponseAsserts;
 
     public function testIndexRendersCorrectly()
@@ -69,5 +70,437 @@ class ReadTest extends \PHPUnit\Framework\TestCase
         ]);
 
         $this->assertResponseCode($response, 302, '/read?page=1');
+    }
+
+    public function testCreateMarksLinksAsReadAndRedirects()
+    {
+        $user = $this->login();
+        $news = $user->news();
+        $read_list = $user->readList();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponse($response, 302, '/news');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $link_to_read_list = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $read_list->id,
+        ]);
+        $this->assertFalse($exists_in_news, 'The link should not be in news.');
+        $this->assertNotNull($link_to_read_list, 'The link should be in read list.');
+    }
+
+    public function testCreateRemovesFromBookmarks()
+    {
+        $user = $this->login();
+        $news = $user->news();
+        $bookmarks = $user->bookmarks();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+        $link_to_bookmarks_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponse($response, 302, '/news');
+        $exists_in_bookmarks = models\LinkToCollection::exists($link_to_bookmarks_id);
+        $this->assertFalse($exists_in_bookmarks, 'The link should not be in bookmarks.');
+    }
+
+    public function testCreateRedirectsToLoginIfNotConnected()
+    {
+        $user_id = $this->create('user');
+        $user = models\User::find($user_id);
+        $news = $user->news();
+        $read_list = $user->readList();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponse($response, 302, '/login?redirect_to=%2Fnews');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $link_to_read_list = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $read_list->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should be in news.');
+        $this->assertNull($link_to_read_list, 'The link should not be in read list.');
+    }
+
+    public function testCreateFailsIfCollectionIsInaccessible()
+    {
+        $user = $this->login();
+        $other_user_id = $this->create('user');
+        $other_user = models\User::find($other_user_id);
+        $news = $other_user->news();
+        $read_list = $user->readList();
+        $link_id = $this->create('link', [
+            'user_id' => $other_user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponse($response, 404);
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $link_to_read_list = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $read_list->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should be in news.');
+        $this->assertNull($link_to_read_list, 'The link should not be in read list.');
+    }
+
+    public function testCreateFailsIfCsrfIsInvalid()
+    {
+        $user = $this->login();
+        $news = $user->news();
+        $read_list = $user->readList();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read", [
+            'csrf' => 'not the token',
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponse($response, 302, '/news');
+        $this->assertFlash('error', 'A security verification failed.');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $link_to_read_list = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $read_list->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should be in news.');
+        $this->assertNull($link_to_read_list, 'The link should not be in read list.');
+    }
+
+    public function testLaterMarksNewsLinksToReadLaterAndRedirects()
+    {
+        $user = $this->login();
+        $bookmarks = $user->bookmarks();
+        $news = $user->news();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/later", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponseCode($response, 302, '/news');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $link_to_bookmarks = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+        $this->assertFalse($exists_in_news, 'The link should no longer be in news.');
+        $this->assertNotNull($link_to_bookmarks, 'The link should be in bookmarks.');
+    }
+
+    public function testLaterJustRemovesFromNewsIfAlreadyBookmarked()
+    {
+        $user = $this->login();
+        $bookmarks = $user->bookmarks();
+        $news = $user->news();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+        $link_to_bookmarks_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/later", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponseCode($response, 302, '/news');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $exists_in_bookmarks = models\LinkToCollection::exists($link_to_bookmarks_id);
+        $this->assertFalse($exists_in_news, 'The link should no longer be in news.');
+        $this->assertTrue($exists_in_bookmarks, 'The link should be in bookmarks.');
+    }
+
+    public function testLaterRedirectsToLoginIfNotConnected()
+    {
+        $user_id = $this->create('user');
+        $user = models\User::find($user_id);
+        $bookmarks = $user->bookmarks();
+        $news = $user->news();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/later", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponseCode($response, 302, '/login?redirect_to=%2Fnews');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $link_to_bookmarks = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should still be in news.');
+        $this->assertNull($link_to_bookmarks, 'The link should not be in bookmarks.');
+    }
+
+    public function testLaterFailsIfCollectionIsInaccessible()
+    {
+        $user = $this->login();
+        $other_user_id = $this->create('user');
+        $other_user = models\User::find($other_user_id);
+        $bookmarks = $user->bookmarks();
+        $news = $other_user->news();
+        $link_id = $this->create('link', [
+            'user_id' => $other_user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/later", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponse($response, 404);
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $link_to_bookmarks = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should still be in news.');
+        $this->assertNull($link_to_bookmarks, 'The link should not be in bookmarks.');
+    }
+
+    public function testLaterFailsIfCsrfIsInvalid()
+    {
+        $user = $this->login();
+        $bookmarks = $user->bookmarks();
+        $news = $user->news();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/later", [
+            'csrf' => 'not the token',
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponse($response, 302, '/news');
+        $this->assertFlash('error', 'A security verification failed.');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $link_to_bookmarks = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should still be in news.');
+        $this->assertNull($link_to_bookmarks, 'The link should not be in bookmarks.');
+    }
+
+    public function testNeverMarksNewsLinksToNeverReadAndRedirects()
+    {
+        $user = $this->login();
+        $bookmarks = $user->bookmarks();
+        $news = $user->news();
+        $never_list = $user->neverList();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+        $link_to_bookmarks_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/never", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponseCode($response, 302, '/news');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $exists_in_bookmarks = models\LinkToCollection::exists($link_to_bookmarks_id);
+        $link_to_never_list = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $never_list->id,
+        ]);
+        $this->assertFalse($exists_in_news, 'The link should no longer be in news.');
+        $this->assertFalse($exists_in_bookmarks, 'The link should no longer be in bookmarks.');
+        $this->assertNotNull($link_to_never_list, 'The link should be in the never list.');
+    }
+
+    public function testNeverRedirectsToLoginIfNotConnected()
+    {
+        $user_id = $this->create('user');
+        $user = models\User::find($user_id);
+        $bookmarks = $user->bookmarks();
+        $news = $user->news();
+        $never_list = $user->neverList();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+        $link_to_bookmarks_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/never", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponseCode($response, 302, '/login?redirect_to=%2Fnews');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $exists_in_bookmarks = models\LinkToCollection::exists($link_to_bookmarks_id);
+        $link_to_never_list = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $never_list->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should be in news.');
+        $this->assertTrue($exists_in_bookmarks, 'The link should be in bookmarks.');
+        $this->assertNull($link_to_never_list, 'The link should not be in the never list.');
+    }
+
+    public function testNeverFailsIfCollectionIsInaccessible()
+    {
+        $user = $this->login();
+        $other_user_id = $this->create('user');
+        $other_user = models\User::find($other_user_id);
+        $bookmarks = $other_user->bookmarks();
+        $news = $other_user->news();
+        $never_list = $user->neverList();
+        $link_id = $this->create('link', [
+            'user_id' => $other_user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+        $link_to_bookmarks_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/never", [
+            'csrf' => $user->csrf,
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponseCode($response, 404);
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $exists_in_bookmarks = models\LinkToCollection::exists($link_to_bookmarks_id);
+        $link_to_never_list = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $never_list->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should be in news.');
+        $this->assertTrue($exists_in_bookmarks, 'The link should be in bookmarks.');
+        $this->assertNull($link_to_never_list, 'The link should not be in the never list.');
+    }
+
+    public function testNeverFailsIfCsrfIsInvalid()
+    {
+        $user = $this->login();
+        $bookmarks = $user->bookmarks();
+        $news = $user->news();
+        $never_list = $user->neverList();
+        $link_id = $this->create('link', [
+            'user_id' => $user->id,
+        ]);
+        $link_to_news_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $news->id,
+        ]);
+        $link_to_bookmarks_id = $this->create('link_to_collection', [
+            'link_id' => $link_id,
+            'collection_id' => $bookmarks->id,
+        ]);
+
+        $response = $this->appRun('post', "/collections/{$news->id}/read/never", [
+            'csrf' => 'not the token',
+            'from' => \Minz\Url::for('news'),
+        ]);
+
+        $this->assertResponseCode($response, 302, '/news');
+        $this->assertFlash('error', 'A security verification failed.');
+        $exists_in_news = models\LinkToCollection::exists($link_to_news_id);
+        $exists_in_bookmarks = models\LinkToCollection::exists($link_to_bookmarks_id);
+        $link_to_never_list = models\LinkToCollection::findBy([
+            'link_id' => $link_id,
+            'collection_id' => $never_list->id,
+        ]);
+        $this->assertTrue($exists_in_news, 'The link should be in news.');
+        $this->assertTrue($exists_in_bookmarks, 'The link should be in bookmarks.');
+        $this->assertNull($link_to_never_list, 'The link should not be in the never list.');
     }
 }
